worker/requirements.txt
pandas>=2.0.0
plotly>=5.0.0
sqlalchemy>=2.0.0
pillow>=10.0.0
google-generativeai>=0.3.0
requests>=2.25.0
schedule>=1.0.0
psycopg2-binary>=2.9.0
kaleido>=0.2.0
numpy>=1.24.0,<2.0.0
pandas-ta-openbb>=0.4.0
pytz>=2021.1
scipy>=1.7.0
anthropic>=0.3.0
click>=8.0.0
psutil>=5.8.0